+++ e2e_tests/tests/experiment/experiment.py
@@ -211,6 +211,21 @@ def experiment_has_completed_workload(experiment_id: int) -> bool:
                 return True
     return False
 
+def create_exp_get_trial_id():
+    exp_id = create_experiment(
+        conf.fixtures_path("no_op/single-one-short-step.yaml"), conf.fixtures_path("no_op")
+    )
+    
+    wait_for_experiment_state(exp_id, determinedexperimentv1State.STATE_COMPLETED)
+    session = test_session()
+    trials = bindings.get_GetExperimentTrials(session, experimentId=exp_id).trials
+
+    assert len(trials) > 0
+    trial_example = trials[0]
+    trial_id = trial_example.id
+    return trial_id
+
+
 
 def test_session() -> session.Session:
     murl = conf.make_master_url()
