--- e2e_tests/tests/trial/__init__.py
@@ -1,1 +1,0 @@
-from .trial import create_exp_get_trial_id, test_session
